wip: even more broken
this moves the Obligation from rustc_infer to rustc_middle in order to
use it as part of OverflowError, so that the query callsites can handle
the OverflowError by terminating error reporting.

this has a compilation error in rustc_middle, which comes from the
inability to derive TypeFoldable and Lift for some of the moved types.
this in itself is likely to be easily fixed, since I don't expect
any/many of the Obligation parts to have much to typefold or lift over;
I may have done a mistake in just trying to derive it all over, but I
was running out of time yesterday.

as the Obligation moved to rustc_middle, I defined a similar type alias
to PredicateObligation: PredicateOverflow, which seemed the most common
return type.

note of course, this doesn't typecheck so perhaps this isn't feasible
approach at all.